fix: sidebar semantics
