+++ src/components/Sidebar.js
@@ -1,8 +1,9 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { useTheme } from 'emotion-theming';
 import isPropValid from '@emotion/is-prop-valid';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faFilter,
   faAngleRight,
@@ -10,12 +11,15 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import { NavLink } from 'react-router-dom';
 
+import { Media } from '../components/Media';
 import VisuallyHidden from './VisuallyHidden';
 import Box from './Box';
 import Text from './Text';
 import Icon from './Icon';
 import LocationSearch from './LocationSearch';
 import Filters from './Filters';
+import Button from '../components/Button';
+import Drawer from '../components/Drawer';
 
 const Arrow = styled((props) => <Icon icon={faAngleRight} {...props} />, {
   shouldForwardProp: (prop) => {
@@ -44,100 +48,174 @@ const Sidebar = ({
   filters,
   onFilterChange,
   onSelectedItemChange,
+  onUpdateMapPosition,
   mapCenter,
 }) => {
   const theme = useTheme();
   const [isFilterExpanded, setIsFilterExpanded] = useState(false);
+  const [isFiltersExpanded, setIsFiltersExpanded] = useState(false);
+  const filterToggleRef = useRef(null);
 
   return (
-    <Box
-      position="absolute"
-      top={3}
-      left={3}
-      p={4}
-      borderRadius={35}
-      bg="white"
-      width="100%"
-      maxHeight={`calc(100% - ${theme.space[4]}px)`}
-      maxWidth={326}
-      overflowY="auto"
-    >
-      <section aria-labelledby="heading-search">
-        <h2 id="heading-search">
-          <b>Search</b>
-        </h2>
+    <section aria-labelledby="heading-search">
+      <Media lessThan="md">
+        <VisuallyHidden>
+          <h2 id="heading-search">Search</h2>
+        </VisuallyHidden>
 
-        <Box mt={3}>
-          <LocationSearch onSelectedItemChange={onSelectedItemChange} />
-        </Box>
-      </section>
+        <Box position="absolute" top={0} left={0} p={3} width="100%">
+          <LocationSearch
+            onSelectedItemChange={(center) => onUpdateMapPosition({ center })}
+          />
 
-      <Box as="section" my={4} aria-labelledby="heading-filters">
-        <h2 id="heading-filters">
-          <VisuallyHidden>Filters</VisuallyHidden>
-        </h2>
-
-        <Box display="flex" justifyContent="space-between">
-          <Box
-            as="button"
-            type="button"
-            display="flex"
-            alignItems="center"
-            aria-expanded={isFilterExpanded}
-            onClick={() => setIsFilterExpanded(!isFilterExpanded)}
-          >
-            <Icon icon={faFilter} fixedWidth size="lg" />
-            <Box mx={2}>
-              <Text lineHeight={1}>
-                <b>Filter</b>
-              </Text>
-            </Box>
-            <Arrow isExpanded={isFilterExpanded} />
+          <Box display="flex" justifyContent="center" mt={3}>
+            <Button
+              ref={filterToggleRef}
+              variant="secondary"
+              icon={<FontAwesomeIcon icon={faFilter} />}
+              aria-expanded={isFiltersExpanded}
+              onClick={() => setIsFiltersExpanded(!isFiltersExpanded)}
+            >
+              Filter Map
+            </Button>
           </Box>
 
-          {isFilterExpanded && (
-            <Text fontSize={12}>
+          <Drawer visible={isFiltersExpanded} animateFrom="left">
+            <Box display="flex" justifyContent="space-between" mb={4}>
+              <Box display="flex" alignItems="flex-end">
+                <FontAwesomeIcon icon={faFilter} fixedWidth size="lg" />
+                <Box as="h2" mx={2}>
+                  <Text lineHeight={1}>
+                    <b>Filter</b>
+                  </Text>
+                </Box>
+              </Box>
+
+              <Text fontSize={12}>
+                <Box
+                  as="button"
+                  type="button"
+                  onClick={() => onFilterChange({})}
+                  border={0}
+                  borderBottom={2}
+                  borderStyle="solid"
+                >
+                  Reset Filter
+                </Box>
+              </Text>
+            </Box>
+
+            <Filters filters={filters} onFilterChange={onFilterChange} />
+
+            <Box display="flex" justifyContent="center" mt={4}>
+              <Button
+                onClick={() => {
+                  setIsFiltersExpanded(false);
+
+                  // return focus to the control that invoked the filter overlay
+                  filterToggleRef.current.focus();
+                }}
+                css={{
+                  width: '100%',
+                }}
+              >
+                Done
+              </Button>
+            </Box>
+          </Drawer>
+        </Box>
+      </Media>
+
+      <Media greaterThan="sm">
+        <Box
+          position="absolute"
+          top={3}
+          left={3}
+          p={4}
+          borderRadius={35}
+          bg="white"
+          width="100%"
+          maxHeight={`calc(100% - ${theme.space[4]}px)`}
+          maxWidth={326}
+          overflowY="auto"
+        >
+          <h2 id="heading-search">
+            <b>Search</b>
+          </h2>
+
+          <Box mt={3}>
+            <LocationSearch onSelectedItemChange={onSelectedItemChange} />
+          </Box>
+
+          <Box as="section" my={4} aria-labelledby="heading-filters">
+            <h2 id="heading-filters">
+              <VisuallyHidden>Filters</VisuallyHidden>
+            </h2>
+
+            <Box display="flex" justifyContent="space-between">
               <Box
                 as="button"
                 type="button"
-                onClick={() => onFilterChange({})}
-                border={0}
-                borderBottom={2}
-                borderStyle="solid"
+                display="flex"
+                alignItems="center"
+                aria-expanded={isFilterExpanded}
+                onClick={() => setIsFilterExpanded(!isFilterExpanded)}
               >
-                Reset Filter
+                <Icon icon={faFilter} fixedWidth size="lg" />
+                <Box mx={2}>
+                  <Text lineHeight={1}>
+                    <b>Filter</b>
+                  </Text>
+                </Box>
+                <Arrow isExpanded={isFilterExpanded} />
               </Box>
-            </Text>
-          )}
-        </Box>
 
-        <Box pt={4} hidden={!isFilterExpanded}>
-          <Filters filters={filters} onFilterChange={onFilterChange} />
-        </Box>
-      </Box>
+              {isFilterExpanded && (
+                <Text fontSize={12}>
+                  <Box
+                    as="button"
+                    type="button"
+                    onClick={() => onFilterChange({})}
+                    border={0}
+                    borderBottom={2}
+                    borderStyle="solid"
+                  >
+                    Reset Filter
+                  </Box>
+                </Text>
+              )}
+            </Box>
 
-      <Box as="section" mt={4} aria-labelledby="heading-add">
-        <h2 id="heading-add">
-          <VisuallyHidden>Add a toilet</VisuallyHidden>
-        </h2>
-        <StyledNavLink
-          to={`/loos/add?lat=${mapCenter.lat}&lng=${mapCenter.lng}`}
-        >
-          <Icon icon={faPlusCircle} fixedWidth size="lg" />
-          <Box mx={2}>
-            <Text lineHeight={1}>
-              <b>Add a Toilet</b>
-            </Text>
+            <Box pt={4} hidden={!isFilterExpanded}>
+              <Filters filters={filters} onFilterChange={onFilterChange} />
+            </Box>
           </Box>
-          <Arrow />
-        </StyledNavLink>
-      </Box>
-    </Box>
+
+          <Box as="section" mt={4} aria-labelledby="heading-add">
+            <h2 id="heading-add">
+              <VisuallyHidden>Add a toilet</VisuallyHidden>
+            </h2>
+            <StyledNavLink
+              to={`/loos/add?lat=${mapCenter.lat}&lng=${mapCenter.lng}`}
+            >
+              <Icon icon={faPlusCircle} fixedWidth size="lg" />
+              <Box mx={2}>
+                <Text lineHeight={1}>
+                  <b>Add a Toilet</b>
+                </Text>
+              </Box>
+              <Arrow />
+            </StyledNavLink>
+          </Box>
+        </Box>
+      </Media>
+    </section>
   );
 };
 
 Sidebar.propTypes = {
   filters: PropTypes.object.isRequired,
+  onUpdateMapPosition: PropTypes.func.isRequired,
   onFilterChange: PropTypes.func.isRequired,
   onSelectedItemChange: PropTypes.func.isRequired,
 };
